добавлен абстрактный класс для фрагмента ленты
