+++ app/src/main/java/ru/techpark/agregator/FeedFragment.java
@@ -15,9 +15,6 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
-import androidx.lifecycle.Observer;
-import androidx.lifecycle.ViewModelProvider;
-import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.bumptech.glide.Glide;
@@ -27,22 +24,24 @@ import java.util.List;
 
 import ru.techpark.agregator.event.Event;
 
-public class FeedFragment extends Fragment {
+public abstract class FeedFragment extends Fragment {
     private static final String SEARCH_STATE = "SEARCH_STATE";
     private static final String SEARCH_QUERY = "SEARCH_QUERY";
     private static final String PAGE = "PAGE";
-    private static final String TAG = "MainFragment";
-    private static FragmentNavigator navigator;
-    private static FeedAdapter adapter;
-    private static ApiViewModel feedViewModel;
-    private ProgressBar loadingProgress;
-    private int pageCounter = 1;
-    private boolean isAllEvents = false;
-    private boolean isSearch = false;
-    private String searchQuery;
-    private EditText searchField;
-    private ImageButton startSearch;
-    private ImageButton exitSearch;
+    protected static final String TAG = "MainFragment";
+    protected boolean isSearch = false;
+    protected String searchQuery;
+    protected int pageCounter = 1;
+    protected boolean isAllEvents = false;
+
+    protected static FragmentNavigator navigator;
+    protected ProgressBar loadingProgress;
+    protected static FeedViewModel feedViewModel;
+    protected static FeedFragment.FeedAdapter adapter;
+
+    protected EditText searchField;
+    protected ImageButton startSearch;
+    protected ImageButton exitSearch;
 
     @Override
     public void onCreate(@Nullable Bundle savedInstanceState) {
@@ -60,68 +59,6 @@ public class FeedFragment extends Fragment {
         outState.putBoolean(SEARCH_STATE, isSearch);
         outState.putString(SEARCH_QUERY, searchQuery);
         outState.putInt(PAGE, pageCounter);
-    }
-
-    @Override
-    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
-        super.onViewCreated(view, savedInstanceState);
-        navigator = ((FragmentNavigator) getActivity());
-        RecyclerView feed = view.findViewById(R.id.list_of_events);
-        loadingProgress = view.findViewById(R.id.loading_progress);
-        adapter = new FeedAdapter();
-        feed.setAdapter(adapter);
-        feed.setLayoutManager(new LinearLayoutManager(view.getContext()));
-        Observer<List<Event>> observer = Events -> {
-            if (Events != null) {
-                hideLoadingProgress();
-                adapter.setEvents(Events);
-            }
-        };
-        feedViewModel = new ViewModelProvider(this).get(ApiViewModel.class);
-        feedViewModel.getEvents()
-                .observe(getViewLifecycleOwner(), observer);
-
-        searchField = view.findViewById(R.id.search_query);
-        Log.d(TAG, "is search " + isSearch);
-        exitSearch = view.findViewById(R.id.exit_search);
-        if (isSearch)
-            exitSearch.setVisibility(View.VISIBLE);
-        exitSearch.setOnClickListener((l) -> {
-            isSearch = false;
-            isAllEvents = false;
-            pageCounter = 1;
-            loadNextPage();
-            exitSearch.setVisibility(View.GONE);
-            searchField.setText("");
-        });
-        startSearch = view.findViewById(R.id.start_search);
-        startSearch.setOnClickListener((l) -> {
-            searchQuery = searchField.getText().toString();
-            if (!searchQuery.equals("")) {
-                pageCounter = 1;
-                isAllEvents = false;
-                isSearch = true;
-                exitSearch.setVisibility(View.VISIBLE);
-                loadNextPage();
-            }
-        });
-    }
-
-    private void hideLoadingProgress() {
-        Log.d(TAG, "progress hidden");
-        loadingProgress.setVisibility(View.GONE);
-    }
-
-    private void showLoadingProgress() {
-        Log.d(TAG, "progress shown");
-        loadingProgress.setVisibility(View.VISIBLE);
-    }
-
-    private void loadNextPage() {
-        showLoadingProgress();
-        if (isSearch)
-            feedViewModel.addSearchNextPage(searchQuery, pageCounter);
-        else feedViewModel.addFeedNextPage(pageCounter);
     }
 
     @Nullable
@@ -148,7 +85,19 @@ public class FeedFragment extends Fragment {
         }
     }
 
-    private class FeedAdapter extends RecyclerView.Adapter<FeedViewHolder> {
+    protected void hideLoadingProgress() {
+        Log.d(TAG, "progress hidden");
+        loadingProgress.setVisibility(View.GONE);
+    }
+
+    protected void showLoadingProgress() {
+        Log.d(TAG, "progress shown");
+        loadingProgress.setVisibility(View.VISIBLE);
+    }
+
+    protected abstract void loadNextPage();
+
+    protected class FeedAdapter extends RecyclerView.Adapter<FeedFragment.FeedViewHolder> {
 
         private List<Event> events = new ArrayList<>();
 
@@ -172,14 +121,14 @@ public class FeedFragment extends Fragment {
 
         @NonNull
         @Override
-        public FeedViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-            return new FeedViewHolder(LayoutInflater.from(parent.getContext())
+        public FeedFragment.FeedViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
+            return new FeedFragment.FeedViewHolder(LayoutInflater.from(parent.getContext())
                     .inflate(R.layout.feed_elem, parent, false));
         }
 
 
         @Override
-        public void onBindViewHolder(@NonNull FeedViewHolder holder, int position) {
+        public void onBindViewHolder(@NonNull FeedFragment.FeedViewHolder holder, int position) {
 
             Event event = events.get(position);
             holder.title.setText(event.getTitle());
